refactor: make a general tableobjectdetector class and balldetector and laserdetector child classes
